This PR contains two changes that significantly improve the performance of `WorkTree.Status()` (from 491 s to 5 s):

- Skip ignored files when walking through the worktree.
  This signigifantly improves the performance of `Status()`:
  In a repository with 3M ignored files `Status` now takes 5 s instead of 160 s.
- Fix loading of `.gitignore` files in nested ignored directories. 
  These were not matched in the current implementation, as that only checks the ignores of the current directory.
  As a side-effect this change also leads to significantly improved performance: in a repository with 3M ignored files `Status` now takes 160 s instead of 491 s.

I have a backport to `master` ready if desired.