@pjbgf 2025-01-15
@silkeh thanks for proposing this PR. Although this may improve the use case of multiple ignored files, it seems to cause a regression when that is not the case. I did a quick benchmark testing a call to `Status()` on a copy of the `go-git` repo, and these changes seem to have a negative impact on both allocation numbers and time per operation:

```
BenchmarkStatus-16 (v5.13.0)      	      48	  21514427 ns/op	19511754 B/op	   69999 allocs/op
BenchmarkStatus-16 (this PR)      	      34	  34658273 ns/op	42274701 B/op	  106194 allocs/op
```

I am struggling for time to pinpoint optimisation areas in the PR, would you be able to optimise it for the standard use case as well?