@pjbgf 2025-10-28
@onee-only rebasing this PR and testing it against the Linux Kernel is still 40% slower, but to your point, it is not as bad as using the current state of the PR.


```
goos: linux
goarch: amd64
pkg: github.com/go-git/go-git/v6
cpu: AMD Ryzen 7 PRO 8840HS w/ Radeon 780M Graphics
                     │ /tmp/main  │               /tmp/ig               │
                     │   sec/op   │   sec/op     vs base                │
Status/osfs.BoundOS-16   7.846 ± 4%   11.150 ± 5%  +42.12% (p=0.000 n=10)

                     │  /tmp/main   │               /tmp/ig               │
                     │     B/op     │     B/op      vs base               │
Status/osfs.BoundOS-16   805.1Mi ± 0%   809.6Mi ± 0%  +0.56% (p=0.001 n=10)

                     │  /tmp/main  │              /tmp/ig               │
                     │  allocs/op  │  allocs/op   vs base               │
Status/osfs.BoundOS-16   15.60M ± 0%   15.61M ± 0%  +0.05% (p=0.001 n=10)
```

The benchmark is using osfs rather than memfs. We probably should add a `BenchmarkStatus` to the code base so that we all use the same going forwards - ideally running across the built-in storages as they will be impacted in different manners.