@onee-only 2026-03-14
Hello @silkeh, If you don't mind, I'd like to take over this PR and help this proposal get merged. Please let me know if you're planning to work on this someday.